requirements.txt
requests
ipykernel
python-dotenv
openai
pydantic
docling
lancedb
streamlit
tiktoken
sentence-transformers
matplotlib